% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altfuel_api.R
\name{count_results}
\alias{count_results}
\title{Count Station Results}
\usage{
count_results(x)
}
\arguments{
\item{x}{A nrel_api object}
}
\value{
A data frame of station count by fuel station type
}
\description{
Count Station Results
}
\examples{
\dontrun{
x <- nearest_stations(MY_API_KEY, location = "Fort Worth, TX", radius = 5)
count_results(x)
}
}
