% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altfuel_api.R
\name{last_updated}
\alias{last_updated}
\title{Retrieve the Date of Last API Data Update.}
\usage{
last_updated(api_key)
}
\arguments{
\item{api_key}{Character. An authorized API key for the NREL API service. API keys can be requested at \url{https://developer.nrel.gov/signup/}}
}
\value{
An object of class POSIXct containing the date and time of the most recent data update.
}
\description{
Retrieve the Date of Last API Data Update.
}
\examples{
\dontrun{
last_updated(MY_API_KEY)
}

}
