% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altfuel_api.R
\name{stations}
\alias{stations}
\title{Extract station data from API object}
\usage{
stations(x)
}
\arguments{
\item{x}{An nrel_api object.}
}
\value{
A data frame with the station listing from the nrel_api object.
}
\description{
Extract station data from API object
}
\examples{
\dontrun{
my_params <- nrel_params(fuel_type = "ELEC", ev_connector_type = "TESLA")
my_stations <- get_all_stations(MY_API_KEY, my_params)
stations(my_stations)
}
}
