% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_params}
\alias{update_params}
\title{Update Non-Null Parameters}
\usage{
update_params(params, updates)
}
\arguments{
\item{params}{list of parameters}

\item{updates}{list of new parameters}
}
\value{
an updated list of parameters
}
\description{
Update Non-Null Parameters
}
\examples{
my_params <- list(name = "Matt Foley")
my_updates <- list(name = NULL, location = "a VAN down by the RIVER")
new_params <- update_params(params = my_params, updates = my_updates)
}
